Authorization in bookingResource methods should be done

+++ src/main/java/nl/utwente/resource/BookingResource.java
@@ -1,8 +1,5 @@
 package nl.utwente.resource;
 
-import com.fasterxml.jackson.databind.node.JsonNodeFactory;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-
 import nl.utwente.dao.BookingDao;
 import nl.utwente.dao.ParticipantDao;
 import nl.utwente.exceptions.BookingException;
@@ -11,23 +8,17 @@ import nl.utwente.model.OutputBooking;
 import nl.utwente.model.SpecifiedBooking;
 import nl.utwente.model.User;
 
-import javax.annotation.security.DeclareRoles;
-import javax.annotation.security.PermitAll;
-import javax.annotation.security.RolesAllowed;
-import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import javax.ws.rs.*;
 import javax.ws.rs.core.*;
 import java.util.List;
-import java.util.Locale;
+import java.util.Objects;
 
-import static nl.utwente.exceptions.ExceptionHandling.throw400;
-import static nl.utwente.exceptions.ExceptionHandling.throw404;
+import static nl.utwente.dao.BookingDao.getEmailOfBookingOwnerFromBookingID;
+import static nl.utwente.exceptions.ExceptionHandling.*;
 
 @Path("/booking")
-@DeclareRoles({"ADMIN","USER"})
-@DenyAll
 public class BookingResource {
     @Context HttpServletResponse response;
     @Context
@@ -42,12 +33,11 @@ public class BookingResource {
     @GET
     @Produces(MediaType.APPLICATION_JSON)
     @Path("/{bookingID}")
-    @RolesAllowed("ADMIN")
     public OutputBooking getSpecificBooking(@PathParam("bookingID") int bookingID) {
         try {
-            return BookingDao.getSpecificBooking(bookingID);
+            return BookingDao.getOutputBooking(bookingID);
         } catch (InvalidBookingIDException e) {
-            throw404(e);
+            throw404(e.getMessage());
         }
 
 
@@ -67,10 +57,24 @@ public class BookingResource {
         try {
             return BookingDao.createBooking(booking);
         } catch (BookingException e) {
-            throw400(e);
+            throw400(e.getMessage());
         }
         return 0;
     }
+
+    @GET
+    @Produces(MediaType.APPLICATION_JSON)
+    @Path("/{bookingID}/participants")
+    public List<User> getParticipants(@PathParam("bookingID") int bookingID) {
+        try {
+            return ParticipantDao.getParticipantsOfBooking(bookingID);
+        } catch (InvalidBookingIDException e) {
+            throw404(e.getMessage());
+        }
+
+        return null;
+    }
+
 
     /**
      * Updates a specified booking.
@@ -85,30 +89,28 @@ public class BookingResource {
         @PathParam("bookingID") int bookingID,
         @Valid SpecifiedBooking booking
     ) {
-        try {
-            if (BookingDao.updateBooking(bookingID, booking)) {
-                return booking;
-            } else {
-                throw new InternalServerErrorException("Something went wrong in updateBooking, but that's all we know");
+        if (securityContext.getUserPrincipal() != null) { // If logged in
+            try {
+                if (Objects.equals(securityContext.getUserPrincipal().getName(),
+                getEmailOfBookingOwnerFromBookingID(bookingID))) { // If owner = logged in user
+
+                    if (BookingDao.updateBooking(bookingID, booking)) {
+                        return booking;
+                    } else {
+                        throw500("Something went wrong in updateBooking, but that's all we know");
+                    }
+                } else {
+                    throw403("You are not authorized to edit this person's booking");
+                }
+
+            } catch (InvalidBookingIDException e) {
+                throw404(e.getMessage());
+            } catch (BookingException e) {
+                throw400(e.getMessage());
             }
-        } catch (InvalidBookingIDException e) {
-            throw404(e);
-        } catch (BookingException e) {
-            throw400(e);
+        } else {
+            throw401("You are not logged in");
         }
-        return null;
-    }
-
-    @GET
-    @Produces(MediaType.APPLICATION_JSON)
-    @Path("/{bookingID}/participants")
-    public List<User> getParticipants(@PathParam("bookingID") int bookingID) {
-        try {
-            return ParticipantDao.getParticipantsOfBooking(bookingID);
-        } catch (InvalidBookingIDException e) {
-            throw404(e);
-        }
-
         return null;
     }
 
@@ -119,16 +121,89 @@ public class BookingResource {
      *         successfully deleted
      */
     @DELETE
-    @RolesAllowed("ADMIN")
     @Produces(MediaType.APPLICATION_JSON)
     @Path("/{bookingID}")
     public void deleteBooking(@PathParam("bookingID") int bookingID) {
-        try {
-            BookingDao.deleteBooking(bookingID);
-        } catch (InvalidBookingIDException e) {
-            throw404(e);
+        if (securityContext.getUserPrincipal() != null) { // If logged in
+            try {
+                if (Objects.equals(securityContext.getUserPrincipal().getName(),
+                    getEmailOfBookingOwnerFromBookingID(bookingID))) { // If owner = logged in user
+                    BookingDao.deleteBooking(bookingID);
+                } else {
+                    throw403("You are not authorized to remove this person's booking");
+                }
+            } catch (InvalidBookingIDException e) {
+                throw404(e.getMessage());
+            }
+        } else {
+            throw401("You are not logged in");
         }
 
 
     }
+
+    /**
+     * Updates a specified booking.
+     * @param bookingID ID specifying the booking to be updated
+     * @return JSON object containing a "success" boolean field specifying whether the booking was
+     *         successfully updated
+     */
+    @PUT
+    @Produces(MediaType.APPLICATION_JSON)
+    @Path("/admin/{bookingID}/update")
+    public SpecifiedBooking adminUpdateBooking(
+        @PathParam("bookingID") int bookingID,
+        @Valid SpecifiedBooking booking
+    ) {
+        if (securityContext.getUserPrincipal() != null) { // If logged in
+            try {
+                if (securityContext.isUserInRole("ADMIN")) {
+
+                    if (BookingDao.updateBooking(bookingID, booking)) {
+                        return booking;
+                    } else {
+                        throw500("Something went wrong in updateBooking, but that's all we know");
+                    }
+                } else {
+                    throw403("You are not authorized to edit this person's booking");
+                }
+
+            } catch (InvalidBookingIDException e) {
+                throw404(e.getMessage());
+            } catch (BookingException e) {
+                throw400(e.getMessage());
+            }
+        } else {
+            throw401("You are not logged in");
+        }
+        return null;
+    }
+
+    /**
+     * Deletes a specific booking
+     * @param bookingID Path parameter specifying the booking to be deleted
+     * @return JSON object containing a "success" boolean field specifying whether the booking was
+     *         successfully deleted
+     */
+    @DELETE
+    @Produces(MediaType.APPLICATION_JSON)
+    @Path("/admin/{bookingID}")
+    public void adminDeleteBooking(@PathParam("bookingID") int bookingID) {
+        if (securityContext.getUserPrincipal() != null) { // If logged in
+            try {
+                if (securityContext.isUserInRole("ADMIN")) {
+                    BookingDao.deleteBooking(bookingID);
+                } else {
+                    throw403("You are not an administrator");
+                }
+            } catch (InvalidBookingIDException e) {
+                throw404(e.getMessage());
+            }
+        } else {
+            throw401("You are not logged in");
+        }
+
+
+    }
+
 }
